Applied single responsibility principle
- File system dialog is now opened in the UI
- Storing data is now performed in the repository
- Configuration Control is no longer dependent on file system
Additional changes:
- Renamed some methods to make them more expressive
- Extracted data patching into its own class since the repository got too large
