+++ Plugin/RLTrainingTimer.vcxproj
@@ -95,6 +95,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="configuration\control\TrainingProgramConfigurationControl.cpp" />
+    <ClCompile Include="configuration\control\TrainingProgramDataPatcher.cpp" />
     <ClCompile Include="configuration\control\TrainingProgramListConfigurationControl.cpp" />
     <ClCompile Include="configuration\control\TrainingProgramRepository.cpp" />
     <ClCompile Include="configuration\control\uuid_generator.cpp" />
@@ -129,6 +130,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="configuration\control\TrainingProgramConfigurationControl.h" />
+    <ClInclude Include="configuration\control\TrainingProgramDataPatcher.h" />
     <ClInclude Include="configuration\control\TrainingProgramListConfigurationControl.h" />
     <ClInclude Include="configuration\control\TrainingProgramRepository.h" />
     <ClInclude Include="configuration\control\uuid_generator.h" />
