+++ Plugin/version.h
@@ -2,7 +2,7 @@
 #define VERSION_MAJOR 1
 #define VERSION_MINOR 1
 #define VERSION_PATCH 0
-#define VERSION_BUILD 753
+#define VERSION_BUILD 762
 
 #define stringify(a) stringify_(a)
 #define stringify_(a) #a
