+++ Test/RLTrainingTimerTest/mocks/ITrainingProgramRepositoryMock.h
@@ -6,7 +6,9 @@
 class ITrainingProgramRepositoryMock : public configuration::ITrainingProgramRepository
 {
 	MOCK_METHOD(void, storeData, (const configuration::TrainingProgramListData& data), (override));
-	MOCK_METHOD(void, storeData, (const configuration::TrainingProgramListData& data, const std::filesystem::path &path), (override));
+	MOCK_METHOD(void, storeData, (const configuration::TrainingProgramListData& data, const std::string& location), (override));
 	MOCK_METHOD(configuration::TrainingProgramListData, restoreData, (), (const override));
-	MOCK_METHOD(configuration::TrainingProgramListData, restoreData, (const std::filesystem::path &path), (const override));
+	MOCK_METHOD(configuration::TrainingProgramListData, restoreData, (const std::string& location), (const override));
+	MOCK_METHOD(void, exportSingleTrainingProgram, (const configuration::TrainingProgramData& data, const std::string& location), (override));
+	MOCK_METHOD(configuration::TrainingProgramData, importSingleTrainingProgram, (const std::string& location), (const override));
 };
